Refactor with new class to select pokemons
