+++ src/main/kotlin/repository/PokemonDictionary.kt
@@ -1,3 +1,5 @@
+package repository
+
 import model.pokemon.*
 
 class PokemonDictionary {
